feat(chat_template): enhance chat templates to have default system messages and support for add_generation_prompt
The llama 3 model now has a default system message if the system message is not found in messages. Also, every chat template now will add a small text at the end of the string to prompt the model to generate response if add_generation_prompt=True in tokenizer.apply_chat_template.